# membership-service/api/membership/v1/membership_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.12
// source: membership/v1/membership.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMembershipGetAssets = "/api.membership.v1.Membership/GetAssets"
const OperationMembershipGetProfile = "/api.membership.v1.Membership/GetProfile"
const OperationMembershipGetReferrals = "/api.membership.v1.Membership/GetReferrals"
const OperationMembershipGetWithdrawals = "/api.membership.v1.Membership/GetWithdrawals"
const OperationMembershipLogin = "/api.membership.v1.Membership/Login"
const OperationMembershipRefreshToken = "/api.membership.v1.Membership/RefreshToken"
const OperationMembershipRegister = "/api.membership.v1.Membership/Register"
const OperationMembershipRequestWithdrawal = "/api.membership.v1.Membership/RequestWithdrawal"
const OperationMembershipUpdateAsset = "/api.membership.v1.Membership/UpdateAsset"
const OperationMembershipUpdateProfile = "/api.membership.v1.Membership/UpdateProfile"

type MembershipHTTPServer interface {
	// GetAssets Asset Management
	GetAssets(context.Context, *GetAssetsRequest) (*GetAssetsReply, error)
	// GetProfile User Profile
	GetProfile(context.Context, *GetProfileRequest) (*GetProfileReply, error)
	// GetReferrals Referral System
	GetReferrals(context.Context, *GetReferralsRequest) (*GetReferralsReply, error)
	GetWithdrawals(context.Context, *GetWithdrawalsRequest) (*GetWithdrawalsReply, error)
	Login(context.Context, *LoginRequest) (*LoginReply, error)
	RefreshToken(context.Context, *RefreshTokenRequest) (*RefreshTokenReply, error)
	// Register User Authentication
	Register(context.Context, *RegisterRequest) (*RegisterReply, error)
	RequestWithdrawal(context.Context, *WithdrawalRequest) (*WithdrawalReply, error)
	UpdateAsset(context.Context, *UpdateAssetRequest) (*UpdateAssetReply, error)
	UpdateProfile(context.Context, *UpdateProfileRequest) (*UpdateProfileReply, error)
}

func RegisterMembershipHTTPServer(s *http.Server, srv MembershipHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/auth/register", _Membership_Register0_HTTP_Handler(srv))
	r.POST("/v1/auth/login", _Membership_Login0_HTTP_Handler(srv))
	r.POST("/v1/auth/refresh", _Membership_RefreshToken0_HTTP_Handler(srv))
	r.GET("/v1/users/profile", _Membership_GetProfile0_HTTP_Handler(srv))
	r.PUT("/v1/users/profile", _Membership_UpdateProfile0_HTTP_Handler(srv))
	r.GET("/v1/users/assets", _Membership_GetAssets0_HTTP_Handler(srv))
	r.PUT("/v1/users/assets/{asset_id}", _Membership_UpdateAsset0_HTTP_Handler(srv))
	r.GET("/v1/users/referrals", _Membership_GetReferrals0_HTTP_Handler(srv))
	r.POST("/v1/users/withdrawals", _Membership_RequestWithdrawal0_HTTP_Handler(srv))
	r.GET("/v1/users/withdrawals", _Membership_GetWithdrawals0_HTTP_Handler(srv))
}

func _Membership_Register0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RegisterRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipRegister)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Register(ctx, req.(*RegisterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RegisterReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_Login0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_RefreshToken0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RefreshTokenRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipRefreshToken)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RefreshToken(ctx, req.(*RefreshTokenRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RefreshTokenReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_GetProfile0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetProfileRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipGetProfile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetProfile(ctx, req.(*GetProfileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetProfileReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_UpdateProfile0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateProfileRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipUpdateProfile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateProfile(ctx, req.(*UpdateProfileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateProfileReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_GetAssets0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetAssetsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipGetAssets)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetAssets(ctx, req.(*GetAssetsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetAssetsReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_UpdateAsset0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateAssetRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipUpdateAsset)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateAsset(ctx, req.(*UpdateAssetRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateAssetReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_GetReferrals0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetReferralsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipGetReferrals)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetReferrals(ctx, req.(*GetReferralsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetReferralsReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_RequestWithdrawal0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in WithdrawalRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipRequestWithdrawal)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RequestWithdrawal(ctx, req.(*WithdrawalRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*WithdrawalReply)
		return ctx.Result(200, reply)
	}
}

func _Membership_GetWithdrawals0_HTTP_Handler(srv MembershipHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetWithdrawalsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipGetWithdrawals)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetWithdrawals(ctx, req.(*GetWithdrawalsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetWithdrawalsReply)
		return ctx.Result(200, reply)
	}
}

type MembershipHTTPClient interface {
	// GetAssets Asset Management
	GetAssets(ctx context.Context, req *GetAssetsRequest, opts ...http.CallOption) (rsp *GetAssetsReply, err error)
	// GetProfile User Profile
	GetProfile(ctx context.Context, req *GetProfileRequest, opts ...http.CallOption) (rsp *GetProfileReply, err error)
	// GetReferrals Referral System
	GetReferrals(ctx context.Context, req *GetReferralsRequest, opts ...http.CallOption) (rsp *GetReferralsReply, err error)
	GetWithdrawals(ctx context.Context, req *GetWithdrawalsRequest, opts ...http.CallOption) (rsp *GetWithdrawalsReply, err error)
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
	RefreshToken(ctx context.Context, req *RefreshTokenRequest, opts ...http.CallOption) (rsp *RefreshTokenReply, err error)
	// Register User Authentication
	Register(ctx context.Context, req *RegisterRequest, opts ...http.CallOption) (rsp *RegisterReply, err error)
	RequestWithdrawal(ctx context.Context, req *WithdrawalRequest, opts ...http.CallOption) (rsp *WithdrawalReply, err error)
	UpdateAsset(ctx context.Context, req *UpdateAssetRequest, opts ...http.CallOption) (rsp *UpdateAssetReply, err error)
	UpdateProfile(ctx context.Context, req *UpdateProfileRequest, opts ...http.CallOption) (rsp *UpdateProfileReply, err error)
}

type MembershipHTTPClientImpl struct {
	cc *http.Client
}

func NewMembershipHTTPClient(client *http.Client) MembershipHTTPClient {
	return &MembershipHTTPClientImpl{client}
}

// GetAssets Asset Management
func (c *MembershipHTTPClientImpl) GetAssets(ctx context.Context, in *GetAssetsRequest, opts ...http.CallOption) (*GetAssetsReply, error) {
	var out GetAssetsReply
	pattern := "/v1/users/assets"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMembershipGetAssets))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetProfile User Profile
func (c *MembershipHTTPClientImpl) GetProfile(ctx context.Context, in *GetProfileRequest, opts ...http.CallOption) (*GetProfileReply, error) {
	var out GetProfileReply
	pattern := "/v1/users/profile"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMembershipGetProfile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetReferrals Referral System
func (c *MembershipHTTPClientImpl) GetReferrals(ctx context.Context, in *GetReferralsRequest, opts ...http.CallOption) (*GetReferralsReply, error) {
	var out GetReferralsReply
	pattern := "/v1/users/referrals"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMembershipGetReferrals))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MembershipHTTPClientImpl) GetWithdrawals(ctx context.Context, in *GetWithdrawalsRequest, opts ...http.CallOption) (*GetWithdrawalsReply, error) {
	var out GetWithdrawalsReply
	pattern := "/v1/users/withdrawals"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMembershipGetWithdrawals))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MembershipHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/v1/auth/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MembershipHTTPClientImpl) RefreshToken(ctx context.Context, in *RefreshTokenRequest, opts ...http.CallOption) (*RefreshTokenReply, error) {
	var out RefreshTokenReply
	pattern := "/v1/auth/refresh"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipRefreshToken))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Register User Authentication
func (c *MembershipHTTPClientImpl) Register(ctx context.Context, in *RegisterRequest, opts ...http.CallOption) (*RegisterReply, error) {
	var out RegisterReply
	pattern := "/v1/auth/register"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipRegister))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MembershipHTTPClientImpl) RequestWithdrawal(ctx context.Context, in *WithdrawalRequest, opts ...http.CallOption) (*WithdrawalReply, error) {
	var out WithdrawalReply
	pattern := "/v1/users/withdrawals"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipRequestWithdrawal))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MembershipHTTPClientImpl) UpdateAsset(ctx context.Context, in *UpdateAssetRequest, opts ...http.CallOption) (*UpdateAssetReply, error) {
	var out UpdateAssetReply
	pattern := "/v1/users/assets/{asset_id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipUpdateAsset))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MembershipHTTPClientImpl) UpdateProfile(ctx context.Context, in *UpdateProfileRequest, opts ...http.CallOption) (*UpdateProfileReply, error) {
	var out UpdateProfileReply
	pattern := "/v1/users/profile"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipUpdateProfile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}